+++ library/src/actions/rawCheck/rawCheckAsync.ts
@@ -1,6 +1,6 @@
 import type { BaseValidationAsync, MaybePromise } from '../../types/index.ts';
 import { _addIssue } from '../../utils/index.ts';
-import type { Context, RawCheckIssue } from './types.ts';
+import type { RawCheckContext, RawCheckIssue } from './types.ts';
 
 /**
  * Raw check action async interface.
@@ -30,7 +30,7 @@ export interface RawCheckActionAsync<TInput>
  */
 // @__NO_SIDE_EFFECTS__
 export function rawCheckAsync<TInput>(
-  action: (context: Context<TInput>) => MaybePromise<void>
+  action: (context: RawCheckContext<TInput>) => MaybePromise<void>
 ): RawCheckActionAsync<TInput> {
   return {
     kind: 'validation',