+++ library/src/actions/rawTransform/index.ts
@@ -1,3 +1,3 @@
 export * from './rawTransform.ts';
 export * from './rawTransformAsync.ts';
-export type { RawTransformIssue } from './types.ts';
+export * from './types.ts';