+++ website/src/routes/playground/index.tsx
@@ -28,6 +28,7 @@ import { useResetSignal } from '~/hooks';
 import { BinIcon, CheckIcon, CopyIcon, PlayIcon, ShareIcon } from '~/icons';
 import { trackEvent } from '~/utils';
 import valibotCode from '../../../../library/dist/index.min.js?url';
+import valibotToJsonSchemaCode from '../../../../packages/to-json-schema/dist/index.min.js?url';
 import editorCode from './editorCode.ts?raw';
 import iframeCode from './iframeCode.js?raw';
 
@@ -223,15 +224,20 @@ export default component$(() => {
 
   return (
     <main
-      class="flex w-full flex-1 flex-col lg:flex-row lg:space-x-5 lg:px-10 lg:py-20 2xl:max-w-[1700px] 2xl:space-x-7 2xl:self-center"
+      class="flex w-full flex-1 flex-col lg:flex-row lg:gap-5 lg:px-10 lg:py-20 2xl:max-w-[1700px] 2xl:gap-7 2xl:self-center"
       window:onMessage$={captureLogs}
       window:onKeyDown$={[preventDefault, handleKeyDown]}
       window:onResize$={resetSideBarWidth}
     >
       <div ref={editorElement} class="flex flex-1 overflow-visible lg:relative">
-        <CodeEditor value={initialCode} model={model} onSave$={saveCode} />
+        <CodeEditor
+          class="lg:rounded-3xl lg:border-[3px] lg:border-slate-200 lg:dark:border-slate-800"
+          value={initialCode}
+          model={model}
+          onSave$={saveCode}
+        />
         <EditorButtons
-          class="!hidden lg:!absolute lg:right-10 lg:top-10 lg:z-10 lg:!flex"
+          class="hidden! lg:absolute! lg:flex! lg:right-10 lg:top-10 lg:z-10"
           model={model}
           executeCode$={executeCode}
         />
@@ -256,7 +262,7 @@ export default component$(() => {
           executeCode$={executeCode}
         />
         <IconButton
-          class="!absolute right-8 top-8 z-10 lg:right-10 lg:top-10"
+          class="absolute! right-8 top-8 z-10 lg:right-10 lg:top-10"
           type="button"
           variant="secondary"
           label="Clear logs"
@@ -301,7 +307,12 @@ export default component$(() => {
           <html>
             <head>
               <script type="importmap">
-                { "imports": { "valibot": "${valibotCode}" } }
+                {
+                  "imports": {
+                    "valibot": "${valibotCode}",
+                    "@valibot/to-json-schema": "${valibotToJsonSchemaCode}"
+                  }
+                }
               </script>
               <script>
                 ${iframeCode}
@@ -370,7 +381,7 @@ const EditorButtons = component$<EditorButtonsProps>(
     });
 
     return (
-      <div class={clsx('flex space-x-6', props.class)}>
+      <div class={clsx('flex gap-6', props.class)}>
         <IconButton
           type="button"
           variant="secondary"